Move service lifecycle and task workers out of the root package
